requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
biopython